完成推广单元接口和实现类的编写

+++ ad-service/ad-sponsor/src/main/java/com/remember/dao/AdUnitRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
  * AdUnit Dao
   */
 public interface AdUnitRepository extends JpaRepository<AdUnit,Long>{
-    AdUnit findByPlanIdAndUnitName(Integer planId,String unitName);
+    AdUnit findByPlanIdAndUnitName(Long planId,String unitName);
 
     List<AdUnit> findAllByUnitStatus(Integer unitStatus);
 }
